Extract segment options into their own module
