+++ library/Strive/Options.hs
@@ -7,6 +7,7 @@ module Strive.Options
   , module Strive.Options.Comments
   , module Strive.Options.Friends
   , module Strive.Options.Kudos
+  , module Strive.Options.Segments
   , module Strive.Options.Streams
   , module Strive.Options.Uploads
   ) where
@@ -18,5 +19,6 @@ import Strive.Options.Clubs
 import Strive.Options.Comments
 import Strive.Options.Friends
 import Strive.Options.Kudos
+import Strive.Options.Segments
 import Strive.Options.Streams
 import Strive.Options.Uploads
