+++ strive.cabal
@@ -69,6 +69,7 @@ library
         Strive.Options.Comments
         Strive.Options.Friends
         Strive.Options.Kudos
+        Strive.Options.Segments
         Strive.Options.Streams
         Strive.Options.Uploads
         Strive.Types
@@ -112,4 +113,3 @@ test-suite hlint
     buildable: True
     default-language: Haskell2010
     hs-source-dirs: test-suite
-
